requirements.txt
openai==0.27.0
python-decouple==3.8
fastapi==0.92.0
python-multipart==0.0.6
requests==2.28.2